Fix Rust 1.56 warnings and move to u8 for auth bitflags

+++ src/auth.rs
@@ -15,7 +15,7 @@ pub struct Authenticator {
 }
 
 bitflags! {
-    pub struct Access: u32 {
+    pub struct Access: u8 {
         const UPLOAD = 1 << 0;
         const DOWNLOAD = 1 << 1;
         const WEB_UI = 1 << 2;
